Move gate data into src/gates/ directory
