+++ file_lists/source_files_no_main
@@ -3,20 +3,6 @@ src/stim/arg_parse.cc
 src/stim/circuit/circuit.cc
 src/stim/circuit/circuit_instruction.cc
 src/stim/circuit/export_qasm.cc
-src/stim/circuit/gate_data.cc
-src/stim/circuit/gate_data_annotations.cc
-src/stim/circuit/gate_data_blocks.cc
-src/stim/circuit/gate_data_collapsing.cc
-src/stim/circuit/gate_data_controlled.cc
-src/stim/circuit/gate_data_hada.cc
-src/stim/circuit/gate_data_heralded.cc
-src/stim/circuit/gate_data_noisy.cc
-src/stim/circuit/gate_data_pair_measure.cc
-src/stim/circuit/gate_data_pauli.cc
-src/stim/circuit/gate_data_period_3.cc
-src/stim/circuit/gate_data_period_4.cc
-src/stim/circuit/gate_data_pp.cc
-src/stim/circuit/gate_data_swaps.cc
 src/stim/circuit/gate_decomposition.cc
 src/stim/circuit/gate_target.cc
 src/stim/cmd/command_analyze_errors.cc
@@ -51,6 +37,20 @@ src/stim/diagram/lattice_map.cc
 src/stim/diagram/timeline/timeline_3d_drawer.cc
 src/stim/diagram/timeline/timeline_ascii_drawer.cc
 src/stim/diagram/timeline/timeline_svg_drawer.cc
+src/stim/gates/gate_data.cc
+src/stim/gates/gate_data_annotations.cc
+src/stim/gates/gate_data_blocks.cc
+src/stim/gates/gate_data_collapsing.cc
+src/stim/gates/gate_data_controlled.cc
+src/stim/gates/gate_data_hada.cc
+src/stim/gates/gate_data_heralded.cc
+src/stim/gates/gate_data_noisy.cc
+src/stim/gates/gate_data_pair_measure.cc
+src/stim/gates/gate_data_pauli.cc
+src/stim/gates/gate_data_period_3.cc
+src/stim/gates/gate_data_period_4.cc
+src/stim/gates/gate_data_pp.cc
+src/stim/gates/gate_data_swaps.cc
 src/stim/gen/circuit_gen_params.cc
 src/stim/gen/gen_color_code.cc
 src/stim/gen/gen_rep_code.cc
