+++ src/stim.h
@@ -7,8 +7,6 @@
 #include "stim/circuit/circuit.h"
 #include "stim/circuit/circuit_instruction.h"
 #include "stim/circuit/export_qasm.h"
-#include "stim/circuit/gate_data.h"
-#include "stim/circuit/gate_data_table.h"
 #include "stim/circuit/gate_decomposition.h"
 #include "stim/circuit/gate_target.h"
 #include "stim/circuit/stabilizer_flow.h"
@@ -45,6 +43,7 @@
 #include "stim/diagram/timeline/timeline_3d_drawer.h"
 #include "stim/diagram/timeline/timeline_ascii_drawer.h"
 #include "stim/diagram/timeline/timeline_svg_drawer.h"
+#include "stim/gates/gate_data.h"
 #include "stim/gen/circuit_gen_params.h"
 #include "stim/gen/gen_color_code.h"
 #include "stim/gen/gen_rep_code.h"
