+++ src/stim/circuit/circuit.cc
@@ -17,7 +17,7 @@
 #include <string>
 #include <utility>
 
-#include "stim/circuit/gate_data.h"
+#include "stim/gates/gate_data.h"
 #include "stim/circuit/gate_target.h"
 #include "stim/str_util.h"
