+++ src/stim/circuit/circuit.h
@@ -28,7 +28,7 @@
 #include <vector>
 
 #include "stim/circuit/circuit_instruction.h"
-#include "stim/circuit/gate_data.h"
+#include "stim/gates/gate_data.h"
 #include "stim/circuit/gate_target.h"
 #include "stim/mem/monotonic_buffer.h"
 #include "stim/mem/span_ref.h"
