+++ src/stim/circuit/circuit_instruction.pybind.cc
@@ -14,7 +14,7 @@
 
 #include "stim/circuit/circuit_instruction.pybind.h"
 
-#include "stim/circuit/gate_data.h"
+#include "stim/gates/gate_data.h"
 #include "stim/circuit/gate_target.pybind.h"
 #include "stim/py/base.pybind.h"
 #include "stim/str_util.h"
