+++ src/stim/circuit/circuit_instruction.pybind.h
@@ -18,7 +18,7 @@
 #include <pybind11/pybind11.h>
 
 #include "stim/circuit/circuit_instruction.h"
-#include "stim/circuit/gate_data.h"
+#include "stim/gates/gate_data.h"
 #include "stim/circuit/gate_target.h"
 
 namespace stim_pybind {
