+++ src/stim/gates/gate_data.h
