+++ src/stim/gates/gate_data.perf.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "stim/circuit/gate_data.h"
+#include "stim/gates/gate_data.h"
 
 #include <iostream>
