+++ src/stim/gates/gate_data.pybind.cc
@@ -14,7 +14,7 @@
 
 #include "stim/circuit/gate_data.pybind.h"
 
-#include "stim/circuit/gate_data.h"
+#include "stim/gates/gate_data.h"
 #include "stim/circuit/stabilizer_flow.h"
 #include "stim/py/base.pybind.h"
 #include "stim/stabilizers/pauli_string.pybind.h"
