+++ src/stim/gates/gate_data_collapsing.cc
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 #include "stim/circuit/circuit.h"
-#include "stim/circuit/gate_data.h"
+#include "stim/gates/gate_data.h"
 
 using namespace stim;
