+++ src/stim/gates/gate_data_period_4.cc
@@ -14,7 +14,7 @@
 
 #include <complex>
 
-#include "stim/circuit/gate_data.h"
+#include "stim/gates/gate_data.h"
 
 using namespace stim;
