+++ src/stim/gates/gate_data_test.py
