+++ src/stim/simulators/measurements_to_detection_events.inl
@@ -14,7 +14,7 @@
 
 #include <cassert>
 
-#include "stim/circuit/gate_data.h"
+#include "stim/gates/gate_data.h"
 #include "stim/io/measure_record_batch_writer.h"
 #include "stim/io/measure_record_reader.h"
 #include "stim/io/stim_data_formats.h"
