+++ src/stim/simulators/vector_simulator.cc
@@ -16,7 +16,7 @@
 
 #include <cassert>
 
-#include "stim/circuit/gate_data.h"
+#include "stim/gates/gate_data.h"
 #include "stim/mem/simd_util.h"
 #include "stim/stabilizers/pauli_string.h"
