+++ src/stim/stabilizers/tableau.inl
@@ -19,7 +19,7 @@
 #include <map>
 #include <random>
 
-#include "stim/circuit/gate_data.h"
+#include "stim/gates/gate_data.h"
 #include "stim/simulators/vector_simulator.h"
 #include "stim/stabilizers/pauli_string.h"
 #include "stim/stabilizers/tableau.h"
